mejoras en tipados para printear enteros y reales

+++ 03-implementacion-de-un-Scanner-en-FLEX/TP3/include/lexemelist.h
@@ -12,6 +12,6 @@
     void lexeme_list_destroy_and_destroy_elements(t_lexeme_list *self);
     void lexeme_list_add(t_lexeme_list* self, char* lexeme);
     void lexeme_list_add_length(t_lexeme_list* self, char* lexeme);
-    int32_t lexeme_list_sum_as_int(t_lexeme_list* self);
+    int64_t lexeme_list_sum_as_int(t_lexeme_list* self);
 
 #endif // !LL_H_
