+++ 03-implementacion-de-un-Scanner-en-FLEX/TP3/src/lexemelist.c
@@ -28,15 +28,15 @@ void lexeme_list_add_length(t_lexeme_list* self, char* lexeme){
 	list_add_sorted(self->elements, (void*)string_duplicate(lexeme), _is_shorter);
 }
 
-int32_t lexeme_list_sum_as_int(t_lexeme_list* self){
+int64_t lexeme_list_sum_as_int(t_lexeme_list* self){
 	void* _partial_sum(void* element, void* element2){
-		int32_t* r_element = (int32_t*)element;
+		int64_t* r_element = (int64_t*)element;
 		char* r_element2 = (char*)element2;
-		int32_t r_element2_value = (int32_t)atoi(r_element2);
+		int64_t r_element2_value = (int64_t)atol(r_element2);
 		*r_element = *r_element + r_element2_value;
 		return (void*)r_element; 
 	}
-	int32_t seed = 0;
+	int64_t seed = 0;
 	seed = *(int32_t*)list_fold(self->elements, (void*)&seed, _partial_sum);
 	return seed;
 }
